// kraken_frontend/src/api/generated/models/HostCertainty.ts
/* tslint:disable */
/* eslint-disable */
/**
 * kraken
 * The core component of kraken-project
 *
 * The version of the OpenAPI document: 0.4.2
 * Contact: git@omikron.dev
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * The certainty of a host
 * @export
 */
export const HostCertainty = {
    Historical: 'Historical',
    SupposedTo: 'SupposedTo',
    Verified: 'Verified'
} as const;
export type HostCertainty = typeof HostCertainty[keyof typeof HostCertainty];


export function HostCertaintyFromJSON(json: any): HostCertainty {
    return HostCertaintyFromJSONTyped(json, false);
}

export function HostCertaintyFromJSONTyped(json: any, ignoreDiscriminator: boolean): HostCertainty {
    return json as HostCertainty;
}

export function HostCertaintyToJSON(value?: HostCertainty | null): any {
    return value as any;
}